Revert "Login tests added"
