+++ src/test/java/ca/qaguru/oranghrmbatch21/pages/LoginPage.java
@@ -1,34 +1,37 @@
 package ca.qaguru.oranghrmbatch21.pages;
 
-import ca.qaguru.oranghrmbatch21.library.PageBase;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 
-public class LoginPage extends PageBase {
-
+public class LoginPage {
+    private WebDriver driver;
     private String idUsername = "txtUsername";
     private String idPassword = "txtPassword";
     private String idLogin = "btnLogin";
     private String idErrMsg = "spanMessage";
 
     public LoginPage(WebDriver driver) {
-        super(driver);
+        this.driver = driver;
     }
 
     public void login(String username, String password, boolean validCredentials, String expErrMsg){
-        setText(By.id(idUsername), username);
-        setText(By.id(idPassword),password);
-        click(By.id(idLogin));
-
+        WebElement txtUsername = driver.findElement(By.id(idUsername));
+        txtUsername.clear();
+        txtUsername.sendKeys(username);
+        WebElement txtPassword = driver.findElement(By.id(idPassword));
+        txtPassword.clear();
+        txtPassword.sendKeys(password);
+        WebElement btnLogin = driver.findElement(By.id(idLogin));
+        btnLogin.click();
         HeaderPage headerPage = new HeaderPage(driver);
         if(validCredentials){
             Assert.assertTrue(headerPage.isMenuVisible());
         }else {
             Assert.assertFalse(headerPage.isMenuVisible());
-            Assert.assertEquals(getText(By.id(idErrMsg)),expErrMsg, "Incorrect Error Message");
+            WebElement lblErrMsg = driver.findElement(By.id(idErrMsg));
+            Assert.assertEquals(lblErrMsg.getText(),expErrMsg, "Incorrect Error Message");
         }
     }
-
 }
