+++ src/test/java/ca/qaguru/oranghrmbatch21/tests/LoginTests.java
@@ -1,6 +1,5 @@
 package ca.qaguru.oranghrmbatch21.tests;
 
-import ca.qaguru.oranghrmbatch21.library.TestBase;
 import ca.qaguru.oranghrmbatch21.pages.LoginPage;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
@@ -10,7 +9,27 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
-public class LoginTests extends TestBase {
+public class LoginTests {
+    private WebDriver driver;
+    private final String BASE_URL = "https://opensource-demo.orangehrmlive.com/";
+
+    @BeforeTest
+    public void beforeTest(){
+        WebDriverManager.chromedriver().setup();
+        WebDriverManager.firefoxdriver().setup();
+    }
+
+    @BeforeMethod
+    public void beforeMethod(){
+        driver = new ChromeDriver();
+        driver.get(BASE_URL);
+
+    }
+
+    @AfterMethod
+    public void afterMethod(){
+        driver.quit();
+    }
 
     @Test
     public void loginTestWithValidCredentials(){
